add QRCode Share System(#4)
